Change SingleStat to use Double so captured statistics from math based simulations do not loose precision, also leave an overloaded int based add method for the code in the SAPP simulation.

+++ src/alifeSim/Scenario/Math/LVManager.java
@@ -42,9 +42,9 @@ public class LVManager
 		predator_prey_rk4();
 		t++;
 		
-		stat_predator_population.addSample((int)predator_population);
+		stat_predator_population.addSample(predator_population);
 		
-		stat_prey_population.addSample((int)prey_population);
+		stat_prey_population.addSample(prey_population);
 		
 	}
